Lots of fixes and a few features
-Claimerators can now allow everyone to use levers, doors, and buttons
but still not break, place or interact with anything else
-Recipes that require water bottles return the bottle
-Melter recipes can have a container other than a bucket
-Fixed M4 machines not giving XP
-Gave amulets a latin description for what they do
-Fixed insulated plutonium and uraniums hardness
-Fixed hoppers not placing things correctly in the melter
-Fixed glowstone requiring  a bucket to be melted into glow glass
Closed issues
#68
#60
#66
#50

+++ src/main/java/minestrapteam/mods/minestrappolation/inventory/slot/SlotSplitter.java
@@ -65,7 +65,7 @@ public class SlotSplitter extends Slot
 			else if (exp < 1.0F)
 			{
 				exp1 = MathHelper.floor_float(amount * exp);
-				if (exp1 < MathHelper.ceiling_float_int(amount * exp) && (float) Math.random() < amount * exp - exp1)
+				if (exp1 < MathHelper.ceiling_float_int((float)amount * exp) && Math.random() < (double)((float)amount * exp - (float)exp1))
 				{
 					++exp1;
 				}
